pyproject.toml
[tool.poetry]
name = "tiralabra"
version = "0.1.0"
description = ""
authors = ["Silja <44512829+Ronttikasa@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
invoke = "^1.7.1"
rich = "^12.4.4"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
coverage = "^6.3.3"
pylint = "^2.13.9"
autopep8 = "^1.6.0"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
